+++ gtfurb/src/main/java/com/example/gtfurb/services/PessoaRelatorioService.java
@@ -22,7 +22,7 @@ public class PessoaRelatorioService {
         return relatorioRepository.findAll();
     }
 
-    public PessoaRelatorio buscarPorId(Long id) {
+    public PessoaRelatorio buscarPorId(Integer id) {
         return relatorioRepository.findById(id)
                 .orElseThrow(() -> new RuntimeException("Não encontrado"));
     }
@@ -50,7 +50,7 @@ public class PessoaRelatorioService {
         return relatorioRepository.save(pessoaRelatorio);
     }
 
-    public PessoaRelatorio atualizar(Long id, String novoTxt, float tmp_tempoRelatorio ) {
+    public PessoaRelatorio atualizar(Integer id, String novoTxt, float tmp_tempoRelatorio ) {
         if (novoTxt == null || novoTxt.isEmpty()) {
             throw new IllegalArgumentException("Texto do relatório é obrigatório.");
         }   
@@ -67,7 +67,7 @@ public class PessoaRelatorioService {
         return relatorioRepository.save(relatorio);
     }
 
-    public void deletar(Long id) {
+    public void deletar(Integer id) {
         relatorioRepository.deleteById(id);
     }
